Move shared stack to common package
