+++ common/shared_stack.py
